requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
django-rest-framework==0.1.0
djoser==2.1.0
django-cors-headers==3.8.0
